requirements.txt
PyMySQL~=1.1.1
requests~=2.32.3
pytz~=2025.2
Django~=4.2.20
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.5.0
pandas~=2.2.3
gunicorn~=21.2.0
mysqlclient~=2.2.5
whitenoise~=6.6.0
python-dotenv~=1.0.1
django-cors-headers~=4.3.1
celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.4.0
redis==4.5.5
croniter==1.3.8

